Rename impl modules to components for component-based impls
