+++ crates/relayer-components-extra/src/batch/mod.rs
@@ -1,4 +1,4 @@
-pub mod impls;
+pub mod components;
 pub mod traits;
 pub mod types;
 pub mod worker;
