+++ crates/relayer-components-extra/src/components/extra.rs
@@ -1,29 +1,30 @@
 use core::marker::PhantomData;
 
-use crate::batch::impls::message_sender::SendMessagesToBatchWorker;
-use crate::batch::types::sink::BatchWorkerSink;
-use crate::relay::impls::auto_relayers::parallel_bidirectional::ParallelBidirectionalRelayer;
-use crate::relay::impls::auto_relayers::parallel_event::ParallelEventSubscriptionRelayer;
-use crate::relay::impls::auto_relayers::parallel_two_way::ParallelTwoWayAutoRelay;
-use crate::relay::impls::packet_relayers::retry::RetryRelayer;
-use crate::std_prelude::*;
-use crate::telemetry::impls::consensus_state::ConsensusStateTelemetryQuerier;
-use crate::telemetry::impls::status::ChainStatusTelemetryQuerier;
-use ibc_relayer_components::relay::impls::client::update::BuildUpdateClientMessages;
-use ibc_relayer_components::relay::impls::message_senders::chain_sender::SendIbcMessagesToChain;
-use ibc_relayer_components::relay::impls::message_senders::update_client::SendIbcMessagesWithUpdateClient;
-use ibc_relayer_components::relay::impls::messages::skip_update_client::SkipUpdateClient;
-use ibc_relayer_components::relay::impls::messages::wait_update_client::WaitUpdateClient;
-use ibc_relayer_components::relay::impls::packet_relayers::ack::base_ack_packet::BaseAckPacketRelayer;
-use ibc_relayer_components::relay::impls::packet_relayers::general::filter_relayer::FilterRelayer;
-use ibc_relayer_components::relay::impls::packet_relayers::general::full_relay::FullCycleRelayer;
-use ibc_relayer_components::relay::impls::packet_relayers::general::lock::LockPacketRelayer;
-use ibc_relayer_components::relay::impls::packet_relayers::general::log::LoggerRelayer;
-use ibc_relayer_components::relay::impls::packet_relayers::receive::base_receive_packet::BaseReceivePacketRelayer;
-use ibc_relayer_components::relay::impls::packet_relayers::receive::skip_received_packet::SkipReceivedPacketRelayer;
-use ibc_relayer_components::relay::impls::packet_relayers::timeout_unordered::timeout_unordered_packet::BaseTimeoutUnorderedPacketRelayer;
+use ibc_relayer_components::relay::components::message_senders::chain_sender::SendIbcMessagesToChain;
+use ibc_relayer_components::relay::components::message_senders::update_client::SendIbcMessagesWithUpdateClient;
+use ibc_relayer_components::relay::components::update_client::build::BuildUpdateClientMessages;
+use ibc_relayer_components::relay::components::update_client::skip::SkipUpdateClient;
+use ibc_relayer_components::relay::components::update_client::wait::WaitUpdateClient;
+use ibc_relayer_components::relay::components::packet_relayers::general::filter_relayer::FilterRelayer;
+use ibc_relayer_components::relay::components::packet_relayers::ack::base_ack_packet::BaseAckPacketRelayer;
+use ibc_relayer_components::relay::components::packet_relayers::general::lock::LockPacketRelayer;
+use ibc_relayer_components::relay::components::packet_relayers::general::full_relay::FullCycleRelayer;
+use ibc_relayer_components::relay::components::packet_relayers::receive::base_receive_packet::BaseReceivePacketRelayer;
+use ibc_relayer_components::relay::components::packet_relayers::general::log::LoggerRelayer;
+use ibc_relayer_components::relay::components::packet_relayers::timeout_unordered::timeout_unordered_packet::BaseTimeoutUnorderedPacketRelayer;
+use ibc_relayer_components::relay::components::packet_relayers::receive::skip_received_packet::SkipReceivedPacketRelayer;
 use ibc_relayer_components::relay::traits::auto_relayer::{RelayMode, BiRelayMode};
 use ibc_relayer_components::relay::traits::ibc_message_sender::MainSink;
+
+use crate::batch::components::message_sender::SendMessagesToBatchWorker;
+use crate::batch::types::sink::BatchWorkerSink;
+use crate::relay::components::auto_relayers::parallel_bidirectional::ParallelBidirectionalRelayer;
+use crate::relay::components::auto_relayers::parallel_event::ParallelEventSubscriptionRelayer;
+use crate::relay::components::auto_relayers::parallel_two_way::ParallelTwoWayAutoRelay;
+use crate::relay::components::packet_relayers::retry::RetryRelayer;
+use crate::std_prelude::*;
+use crate::telemetry::components::consensus_state::ConsensusStateTelemetryQuerier;
+use crate::telemetry::components::status::ChainStatusTelemetryQuerier;
 
 pub struct ExtraComponents<BaseComponents>(pub PhantomData<BaseComponents>);
