+++ crates/relayer-components-extra/src/relay/mod.rs
@@ -1,1 +1,1 @@
-pub mod impls;
+pub mod components;
