+++ crates/relayer-components-extra/src/telemetry/mod.rs
@@ -1,2 +1,2 @@
-pub mod impls;
+pub mod components;
 pub mod traits;
