+++ crates/relayer-components/src/components/default.rs
@@ -1,20 +1,20 @@
 use core::marker::PhantomData;
 
-use crate::relay::impls::auto_relayers::concurrent_bidirectional::ConcurrentBidirectionalRelayer;
-use crate::relay::impls::auto_relayers::concurrent_event::ConcurrentEventSubscriptionRelayer;
-use crate::relay::impls::client::update::BuildUpdateClientMessages;
-use crate::relay::impls::message_senders::chain_sender::SendIbcMessagesToChain;
-use crate::relay::impls::message_senders::update_client::SendIbcMessagesWithUpdateClient;
-use crate::relay::impls::messages::skip_update_client::SkipUpdateClient;
-use crate::relay::impls::messages::wait_update_client::WaitUpdateClient;
-use crate::relay::impls::packet_relayers::ack::base_ack_packet::BaseAckPacketRelayer;
-use crate::relay::impls::packet_relayers::general::filter_relayer::FilterRelayer;
-use crate::relay::impls::packet_relayers::general::full_relay::FullCycleRelayer;
-use crate::relay::impls::packet_relayers::general::lock::LockPacketRelayer;
-use crate::relay::impls::packet_relayers::general::log::LoggerRelayer;
-use crate::relay::impls::packet_relayers::receive::base_receive_packet::BaseReceivePacketRelayer;
-use crate::relay::impls::packet_relayers::receive::skip_received_packet::SkipReceivedPacketRelayer;
-use crate::relay::impls::packet_relayers::timeout_unordered::timeout_unordered_packet::BaseTimeoutUnorderedPacketRelayer;
+use crate::relay::components::auto_relayers::concurrent_bidirectional::ConcurrentBidirectionalRelayer;
+use crate::relay::components::auto_relayers::concurrent_event::ConcurrentEventSubscriptionRelayer;
+use crate::relay::components::message_senders::chain_sender::SendIbcMessagesToChain;
+use crate::relay::components::message_senders::update_client::SendIbcMessagesWithUpdateClient;
+use crate::relay::components::packet_relayers::ack::base_ack_packet::BaseAckPacketRelayer;
+use crate::relay::components::packet_relayers::general::filter_relayer::FilterRelayer;
+use crate::relay::components::packet_relayers::general::full_relay::FullCycleRelayer;
+use crate::relay::components::packet_relayers::general::lock::LockPacketRelayer;
+use crate::relay::components::packet_relayers::general::log::LoggerRelayer;
+use crate::relay::components::packet_relayers::receive::base_receive_packet::BaseReceivePacketRelayer;
+use crate::relay::components::packet_relayers::receive::skip_received_packet::SkipReceivedPacketRelayer;
+use crate::relay::components::packet_relayers::timeout_unordered::timeout_unordered_packet::BaseTimeoutUnorderedPacketRelayer;
+use crate::relay::components::update_client::build::BuildUpdateClientMessages;
+use crate::relay::components::update_client::skip::SkipUpdateClient;
+use crate::relay::components::update_client::wait::WaitUpdateClient;
 use crate::relay::traits::auto_relayer::RelayMode;
 use crate::relay::traits::ibc_message_sender::MainSink;
 use crate::std_prelude::*;
