+++ crates/relayer-components/src/relay/components/mod.rs
@@ -1,0 +1,4 @@
+pub mod auto_relayers;
+pub mod message_senders;
+pub mod packet_relayers;
+pub mod update_client;
