+++ crates/relayer-components/src/relay/impls/client/mod.rs
@@ -1,2 +1,1 @@
 pub mod create;
-pub mod update;
