--- crates/relayer-components/src/relay/impls/messages/mod.rs
@@ -1,2 +1,0 @@
-pub mod skip_update_client;
-pub mod wait_update_client;
