+++ crates/relayer-components/src/relay/impls/mod.rs
@@ -1,11 +1,7 @@
-pub mod auto_relayers;
 pub mod channel;
 pub mod client;
 pub mod connection;
 pub mod event_relayers;
-pub mod message_senders;
-pub mod messages;
 pub mod packet_clear;
 pub mod packet_filters;
-pub mod packet_relayers;
 pub mod update_client;
