+++ crates/relayer-components/src/relay/mod.rs
@@ -29,6 +29,7 @@
    chains.
 */
 
+pub mod components;
 pub mod impls;
 pub mod traits;
 pub mod types;
